third_party/blink/web_tests/external/wpt/service-workers/cache-storage/cache-keys-attributes-for-service-worker.https-expected.txt

This is a testharness.js-based test.
FAIL Request.IsReloadNavigation should persist. assert_equals: expected "original: false, stored: false" but got "original: undefined, stored: undefined"
FAIL Request.IsHistoryNavigation should persist. assert_equals: expected "original: true, stored: true" but got "original: true, stored: false"
Harness: the test ran to completion.
